// frontend/src/api/model/concourseTagRead.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Qualis API
 * OpenAPI spec version: 0.1.0
 */
import type { ConcourseTagReadColor } from './concourseTagReadColor';

/**
 * Schema for reading a tag.
 */
export interface ConcourseTagRead {
    /**
     * @minLength 1
     * @maxLength 100
     */
    name: string;
    color?: ConcourseTagReadColor;
    id: number;
    project_id: number;
}